+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 =====
 
 * restoreIndex and restoreIndexes in collection will restore previously deleted indexes
+* added max_conflict_retries to handle arango's 1200
+* added single session so AikidoSessio.Holders can share a single request session
 
 1.3.4
 =====